chore(fj-lax): clear all-target clippy test lint debt (98eoz)
Mechanical lint fixes in #[cfg(test)] code, no behavior change (test inputs finite):
- unnecessary_cast (arithmetic.rs): lo/hi already f64 in the f32 clamp ref
- manual_clamp (nn.rs): relu6/hard_tanh refs use .clamp() instead of .max().min()
- type_complexity (tensor_ops.rs): ReplicateCase type alias for the broadcast cases
- cloned_ref_to_slice_refs (tensor_ops.rs): std::slice::from_ref(&v) for argsort
  single-input slices (u32/u64), avoiding the clone

Verified: rch cargo clippy -p fj-lax --all-targets -- -D warnings finished clean.

+++ crates/fj-lax/src/arithmetic.rs
@@ -20563,7 +20563,7 @@ mod tests {
             .collect();
         let wfb: Vec<u32> = xf
             .iter()
-            .map(|&xv| ref_clamp(lo as f64, f64::from(xv), hi as f64) as f32)
+            .map(|&xv| ref_clamp(lo, f64::from(xv), hi) as f32)
             .map(|v| v.to_bits())
             .collect();
         assert!(gfb == wfb, "threaded f32 clamp != serial");
